% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/agnosticImport-methods.R
\docType{methods}
\name{agnosticImport}
\alias{agnosticImport}
\alias{agnosticImport,BigWigFile,GenomicRanges-method}
\alias{agnosticImport,RleList,GenomicRanges-method}
\title{Advanced usage: Import chunks of a genome-wide vector}
\usage{
agnosticImport(signal, sites)

\S4method{agnosticImport}{BigWigFile,GenomicRanges}(signal, sites)

\S4method{agnosticImport}{RleList,GenomicRanges}(signal, sites)
}
\arguments{
\item{signal}{BigWigFile or RleList: Genomic signal}

\item{sites}{GenomicRanges: Regions to be extracted}
}
\value{
NumericList
}
\description{
Import regions from coverage stored as a BigWigFile or RleList and return a NumericList.
}
\examples{
data("CAGE_clusters")
data("CAGE_plus")

# Import from an RleList
agnosticImport(signal=CAGE_plus$WT, sites=subset(CAGE_clusters, strand=="+"))

# Import from a BigWig
\dontrun{
}
}
