% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CAGE_clusters}
\alias{CAGE_clusters}
\alias{CAGE_plus}
\alias{CAGE_minus}
\title{Example data}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
CAGE_clusters

CAGE_plus

CAGE_minus
}
\description{
Included datasets:
}
\keyword{datasets}
