% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_language.R
\name{get_language}
\alias{get_language}
\title{Get language}
\usage{
get_language()
}
\arguments{
\item{language}{- a string to set the language for the downloaded
tables. Options: "ro" - for Romanian and "en" - for English. If no parameter
is given, implicitly downloads tables in Romanian}
}
\description{
Gets the language used for downloading a table from TEMPO ONLINE.
}
