% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-interval.R
\name{intv_bounds}
\alias{intv_bounds}
\alias{intv_start}
\alias{intv_end}
\title{Temporal interval bounds}
\usage{
intv_start(x)

intv_end(x)
}
\arguments{
\item{x}{A temporal interval vector (see \code{\link[=interval]{interval()}}).}
}
\value{
A numeric vector the same length as \code{x}.
}
\description{
\code{intv_start()} and \code{intv_end()} extract the bounds of a temporal interval
vector. The \emph{start} is the chronologically earliest point of each interval; the
\emph{end} is the chronologically latest point.
}
\examples{
x <- interval(c(10, 20), c(30, 40))
intv_start(x)
intv_end(x)
}
