% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intv_misc.R
\name{intv_duration}
\alias{intv_duration}
\alias{intv_seq}
\alias{seq.tempo_interval}
\title{Interval durations}
\usage{
intv_duration(x)

intv_seq(x, ...)

\method{seq}{tempo_interval}(x, ...)
}
\arguments{
\item{x}{A temporal interval vector (see \code{\link[=interval]{interval()}}).}

\item{...}{Passed to \code{\link[=seq]{seq()}}.}
}
\value{
\code{intv_duration()} returns a numeric vector of durations.
\code{intv_seq()} returns a list of numeric vectors, one per interval.
}
\description{
\code{intv_duration()} calculates the duration of each temporal interval.
\code{intv_seq()} generates sequences of values within each temporal interval.
}
\examples{
x <- interval(c(10, 20), c(30, 40))
intv_duration(x)
intv_seq(x, by = 5)
}
