% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intv_similarity.R
\name{intv_similarity}
\alias{intv_similarity}
\alias{intv_delay}
\title{Temporal similarity index}
\usage{
intv_similarity(x, y = NULL, weights = c(0.4, 0.2, 0.4))

intv_delay(x, y)
}
\arguments{
\item{x, y}{Temporal intervals, see \code{\link[=interval]{interval()}}. If \code{y} is \code{NULL} (the
default, calculates pairwise comparisons for each element of \code{x}.}

\item{weights}{Vector of weighting coefficients. Defaults to
\code{c(0.4, 0.2, 0.4)} as suggested by Binding (2010).}
}
\description{
Computes Binding's (2010) index of similarity between pairs of temporal
intervals.
}
\references{
Binding, C. (2010). Implementing Archaeological Time Periods Using CIDOC CRM and SKOS. In: Aroyo, L., et al. The Semantic Web: Research and Applications. ESWC 2010. Lecture Notes in Computer Science, vol 6088. Springer, Berlin, Heidelberg.
}
\keyword{internal}
