% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/quantileTrim-methods.R
\docType{methods}
\name{quantileTrim}
\alias{quantileTrim}
\alias{quantileTrim,matrix,NULL_OR_missing-method}
\alias{quantileTrim,matrix,matrix-method}
\title{Advanced usage: Quantile trim a matrix or matrix pair}
\usage{
quantileTrim(mat1, mat2 = NULL, lower = 0, upper = 1)

\S4method{quantileTrim}{matrix,NULL_OR_missing}(mat1, mat2 = NULL,
  lower = 0, upper = 1)

\S4method{quantileTrim}{matrix,matrix}(mat1, mat2 = NULL, lower = 0,
  upper = 1)
}
\arguments{
\item{mat1}{matrix}

\item{mat2}{matrix or NULL}

\item{lower}{numeric}

\item{upper}{numeric}
}
\value{
trimmed matrix or list of trimmed matrices
}
\description{
Advanced usage: Quantile trim a matrix or matrix pair
}
\examples{
# Make some random matrices
m1 <- replicate(5, rnorm(10))
m2 <- replicate(5, rnorm(10))

# Trim a single matrix
quantileTrim(mat1=m1, upper=0.75)

# Trim a pair of matrices (e.g. forward/reverse strand)
quantileTrim(mat1=m1, mat2=m2, upper=0.75)
}
