% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_language.R
\name{set_language}
\alias{set_language}
\title{Set language}
\usage{
set_language(language = c("ro", "en"))
}
\arguments{
\item{language}{- a string to set the language for the downloaded
tables. Options: "ro" - for Romanian and "en" - for English. If no parameter
is given, implicitly downloads tables in Romanian}
}
\description{
Sets the language used for downloading a table from TEMPO ONLINE.
}
