% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPO-package.R
\docType{package}
\name{TEMPO-package}
\alias{TEMPO}
\alias{TEMPO-package}
\title{TEMPO: Access Romanian National Institute of Statistics Data}
\description{
List and download tables from the Romanian National Institute of Statistics
TEMPO Online database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RProjectRomania/TEMPO}
  \item Report bugs at \url{https://github.com/RProjectRomania/TEMPO/issues}
}

}
\author{
\strong{Maintainer}: Marian Necula \email{marian.necula00@gmail.com}

Authors:
\itemize{
  \item Bogdan Oancea \email{bogdan.oancea@gmail.com}
  \item Ana Maria Tiru \email{tiru.ana21@gmail.com}
}

}
\keyword{internal}
