% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_filter.R
\name{tempo_filter}
\alias{tempo_filter}
\title{tempo_filter}
\usage{
tempo_filter(codes = NULL)
}
\arguments{
\item{codes}{- a list containing one or more strings containing the code for
the table/matrix in TEMPO Online database. See more on
how to obtain a matrix code from \code{\link{tempo_toc}}}
}
\value{
Returns a list containing string vectors.
}
\description{
Get names for subsetting a Tempo Online marix
}
\examples{
tempo_filter("AGR111A")

}
