% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_plot.R
\name{tempo_geo}
\alias{tempo_geo}
\title{Visualize Geospatial Data}
\usage{
tempo_geo(matrix, year, area)
}
\arguments{
\item{matrix}{- the table/matrix downloaded from TEMPO Online database
containing Geospatial Data}

\item{year}{- a string containing the year used to plot Geospatial Data}

\item{area}{- a string to set the type of area (counties, regions, macroregions)
used to group Geospatial Data}
}
\value{
Returns a ggplot object.
}
\description{
tempo_plot creates a map for visualize Geospatial Data 
downloaded from TEMPO Online database
}
\examples{
tempo_plot(matrix=AGR111A, year="2000", area="counties")

}
