% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_geo_codes.R
\name{tempo_geo_codes}
\alias{tempo_geo_codes}
\title{Tempo Geo Codes}
\usage{
tempo_geo_codes(language = c("ro", "en"))
}
\arguments{
\item{language}{- a string to set the language for the downloaded
tables. Options: "ro" - for Romanian and "en" - for English. If no parameter
is given, implicitly downloads tables in Romanian.}
}
\value{
Returns a dataframe object.
}
\description{
tempo_geo_codes 
The function returns a dataframe containing names
and codes for macroregions, regions and counties.
}
\details{

}
\examples{
\dontrun{
tempo_geo_codes(language = "en")
}

}
