% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_get.R
\name{tempo_get}
\alias{tempo_get}
\title{Download a table from TEMPO Online database}
\usage{
tempo_get(matrix_code = NULL, language = NULL, clean = FALSE)
}
\arguments{
\item{matrix_code}{- string containing the code for 
the table/matrix in TEMPO Online database. See more on
how to obtain a matrix code from \code{\link{tempo_toc}}}

\item{language}{- a string to set the language for the downloaded 
tables. Options: "ro" - for Romanian and "en" - for English. If no parameter
is given, implicitly downloads tables in Romanian}

\item{clean}{- calls \code{\link{tempo_clean}} function. Implicitly is set as FALSE}
}
\value{
Returns a dataframe object, which contains a comment attribute with
the full name of the table.
}
\description{
tempo_get downloads a a table from 
TEMPO Online database
}
\details{
This function sends GET/POST requests to TEMPO Online JSON service
and parses the content of the responses, using  \code{\link{httr}} wrappers
for \code{\link{RCurl::httpGET}} and \code{\link{RCurl::httpPOST}}.
The content of the \code{\link{httr::POST}} requests is parsed into a dataframe.
}
\examples{
{tempo_get("SOM101D", "ro")}
}
