% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_last_up.R
\name{tempo_last_up}
\alias{tempo_last_up}
\title{TEMPO Last Updates}
\usage{
tempo_last_up(code = NULL, simplified = FALSE, language = c("ro",
  "en"))
}
\arguments{
\item{code}{String representing the code for
the table/matrix in TEMPO Online database. See more on
how to obtain a matrix code from \code{\link{tempo_toc}}.}

\item{simplified}{Logical indicating whether results 
should be simplified. The default value is FALSE.}

\item{language}{String to set the language for the downloaded
tables. Options: "ro" - for Romanian and "en" - for English. If no parameter
is given, implicitly downloads tables in Romanian.}
}
\value{
Returns a dataframe object.
}
\description{
Returns a dataframe containing the last updates operated on TEMPO tables.
}
\details{

}
\examples{
\dontrun{
tempo_last_up(code = "POP108B", language = "en")
}

}
