% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_options.R
\name{tempo_options}
\alias{tempo_options}
\title{tempo_options}
\usage{
tempo_options(code = NULL)
}
\arguments{
\item{codes}{- a list containing one or more strings containing the code for
the table/matrix in TEMPO Online database. See more on
how to obtain a matrix code from \code{\link{tempo_toc}}}
}
\value{
Returns a list containing named vectors.
}
\description{
Get options for a Tempo Online marix
}
\details{
This functions sends one GET request and parses the content into 
a list of named vector containing options for subsetting a TEMPO Online matrix.
}
\examples{
tempo_options("AGR111A")

}
