% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_search.R
\name{tempo_search}
\alias{tempo_search}
\title{Get TOC for Tempo Online database filtered by keyword}
\usage{
tempo_search(keyword = c())
}
\arguments{
\item{keyword}{Character vector with multiple keywords to be matched.}

\item{language}{String to set the language for the downloaded
tables. Options: "ro" - for Romanian and "en" - for English. If no parameter
is given, implicitly downloads tables in Romanian.}
}
\value{
Returns a dataframe object.
}
\description{
Searches for datasets codes in TOC based on specified keywords, 
to be used with tempo_bulk function.
}
\details{
Downloads Table Of Contents (TOC) for Tempo Online database  
and returns a TOC subset based on specified keywords.
E.g. all TOC entries related to education.
}
\examples{
\dontrun{
tempo_search(c("education", "industry"))
}
}
