% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempotime2date.R
\name{tempotime2date}
\alias{tempotime2date}
\title{Date Conversion from TEMPO Time Format}
\usage{
tempotime2date(matrix, matrix_code)
}
\arguments{
\item{matrix}{- the R dataframe object with time information
in TEMPO time format, representing the table/matrix 
downloaded from TEMPO Online database.}

\item{matrix_code}{- string containing the code for 
the table/matrix in TEMPO Online database. See more on
how to obtain a matrix code from \code{\link{tempo_codes}}}
}
\value{
Returns a R dataframe object.
}
\description{
A function to convert time values from a table
downloaded from TEMPO Online database to objects of class 
\code{\link{Date}} representing calendar dates.
}
\details{
This function converts a column of class character
representing time periods (month, quarter) to a column 
of class \code{\link{Date}} representing calendar dates.
}
\examples{
tempotime2date(AMG157G, "AMG157G")

}
