% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tidyMetaProfile-methods.R
\docType{methods}
\name{tidyMetaProfile}
\alias{tidyMetaProfile}
\alias{tidyMetaProfile,GenomicRanges,BigWigFile_OR_RleList,NULL_OR_missing-method}
\alias{tidyMetaProfile,GenomicRanges,BigWigFile_OR_RleList,BigWigFile_OR_RleList-method}
\alias{tidyMetaProfile,GenomicRangesList,BigWigFile_OR_RleList,NULL_OR_missing-method}
\alias{tidyMetaProfile,GenomicRangesList,BigWigFile_OR_RleList,BigWigFile_OR_RleList-method}
\alias{tidyMetaProfile,GenomicRanges_OR_GenomicRangesList,BigWigFileList_OR_RleListList,NULL_OR_missing-method}
\alias{tidyMetaProfile,GenomicRanges_OR_GenomicRangesList,BigWigFileList_OR_RleListList,BigWigFileList_OR_RleListList-method}
\title{TeMPO: Tidy Meta Profiles}
\usage{
tidyMetaProfile(sites, forward, reverse = NULL, upstream = 100,
  downstream = 100, trimLower = 0, trimUpper = 1,
  sumFun = matrixStats::colMeans2)


  \S4method{tidyMetaProfile}{GenomicRanges,BigWigFile_OR_RleList,NULL_OR_missing}(sites,
  forward, reverse = NULL, upstream = 100, downstream = 100,
  trimLower = 0, trimUpper = 1, sumFun = matrixStats::colMeans2)


  \S4method{tidyMetaProfile}{GenomicRanges,BigWigFile_OR_RleList,BigWigFile_OR_RleList}(sites,
  forward, reverse = NULL, upstream = 100, downstream = 100,
  trimLower = 0, trimUpper = 1, sumFun = matrixStats::colMeans2)


  \S4method{tidyMetaProfile}{GenomicRangesList,BigWigFile_OR_RleList,NULL_OR_missing}(sites,
  forward, reverse = NULL, upstream = 100, downstream = 100,
  trimLower = 0, trimUpper = 1, sumFun = matrixStats::colMeans2)


  \S4method{tidyMetaProfile}{GenomicRangesList,BigWigFile_OR_RleList,BigWigFile_OR_RleList}(sites,
  forward, reverse = NULL, upstream = 100, downstream = 100,
  trimLower = 0, trimUpper = 1, sumFun = matrixStats::colMeans2)


  \S4method{tidyMetaProfile}{GenomicRanges_OR_GenomicRangesList,BigWigFileList_OR_RleListList,NULL_OR_missing}(sites,
  forward, reverse = NULL, upstream = 100, downstream = 100,
  trimLower = 0, trimUpper = 1, sumFun = matrixStats::colMeans2)


  \S4method{tidyMetaProfile}{GenomicRanges_OR_GenomicRangesList,BigWigFileList_OR_RleListList,BigWigFileList_OR_RleListList}(sites,
  forward, reverse = NULL, upstream = 100, downstream = 100,
  trimLower = 0, trimUpper = 1, sumFun = matrixStats::colMeans2)
}
\arguments{
\item{sites}{GenomicRanges: Single-bp genomic locations to calculate the meta-profile over.}

\item{forward}{BigWigFile or RleList: Genome-wide signal stored either on disk as a BigWig-file or in memory as an RleList-object.}

\item{reverse}{BigWigFile, RleList or NULL: If reverse=NULL, the forward signal is taken as unstranded. If not the genomic signal is taken to be stranded. The class of reverse must be the same as forward.}

\item{upstream}{integer: Number of bases to extend upstream.}

\item{downstream}{integer: Number of bases to extend downstream}

\item{trimLower}{numeric: Lower quantile used for trimming.}

\item{trimUpper}{numeric: Upper quantile used for trimming.}

\item{sumFun}{function: The function used for summarising the metaprofile. Should be similar to colSums.}
}
\value{
A tibble with the following columns
\describe{
  \item{pos0}{0-based position relative to sites, useful for plotting}
  \item{pos1}{1-based position relative to sites}
  \item{sense}{Summarized signal for each position}
}
Depending on the input, addditional columns might be added
\describe{
  \item{anti}{If reverse is supplied the genome-wide signal is assumed to be stranded, and the output is split into sense/antisense relative to the orientation of sites.}
  \item{sites}{If sites is a GRangesList, indicates the different sets of sites}
  \item{signal}{if forward/reverse is a BigWigFileList or an RleListList, indicates the different genomic signals}
}
}
\description{
TeMPO: Tidy Meta Profiles
}
