% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/wideMetaProfile-methods.R
\docType{methods}
\name{wideMetaProfile}
\alias{wideMetaProfile}
\alias{wideMetaProfile,GenomicRanges,BigWigFile_OR_RleList,NULL_OR_missing-method}
\alias{wideMetaProfile,GenomicRanges,BigWigFile_OR_RleList,BigWigFile_OR_RleList-method}
\title{Advanced usage: Meta Profile in wide format}
\usage{
wideMetaProfile(sites, forward, reverse = NULL, upstream = 100,
  downstream = 100)


  \S4method{wideMetaProfile}{GenomicRanges,BigWigFile_OR_RleList,NULL_OR_missing}(sites,
  forward, reverse = NULL, upstream = 100, downstream = 100)


  \S4method{wideMetaProfile}{GenomicRanges,BigWigFile_OR_RleList,BigWigFile_OR_RleList}(sites,
  forward, reverse = NULL, upstream = 100, downstream = 100)
}
\arguments{
\item{sites}{GenomicRanges: Single-bp genomic locations to calculate the
meta-profile over.}

\item{forward}{BigWigFile or RleList: Genome-wide signal stored either on
disk as a BigWig-file or in memory as an RleList-object.}

\item{reverse}{BigWigFile, RleList or NULL: If reverse=NULL, the forward
signal is taken as unstranded. If not the genomic signal is taken to be
stranded. The class of reverse must be the same as forward.}

\item{upstream}{integer: Number of bases to extend upstream.}

\item{downstream}{integer: Number of bases to extend downstream}
}
\value{
if reverse=NULL a matrix, else a list of two matrices with names
  "sense" and "anti"
}
\description{
Calculates the wide-format meta profile, with regions in rows and positions
in columns.
}
\examples{
data("CAGE_clusters")
data("CAGE_plus")
data("CAGE_minus")

enhancers <- subset(CAGE_clusters, clusterType == "enhancer")

# Unstranded data gives a single matrix:
wideMetaProfile(sites=enhancers,
                forward=CAGE_plus$WT)

# Stranded data gives two matrices (sense and anti):
wideMetaProfile(sites=enhancers,
                forward=CAGE_plus$WT,
                reverse=CAGE_minus$WT)
}
