% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-interval.R
\name{interval}
\alias{interval}
\alias{is_interval}
\title{Vectors of temporal intervals}
\usage{
interval(start = numeric(), end = numeric())

is_interval(x)
}
\arguments{
\item{start, end}{Numeric vectors specifying the earliest and latest points
in each interval, respectively. Unequal length vectors are recycled to
their common length (if possible) using the rules described in
\code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}.}

\item{x}{Object to test.}
}
\value{
\code{interval()} returns an object with the S3 class \code{"tempo_interval"},
representing a vector of temporal intervals.

\code{is_interval()} returns a logical value.
}
\description{
The \code{interval} class represents the interval between two points in time.
}
\examples{
interval(c(10, 20, 30), c(20, 30, 40))
}
