% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/quantileTrim-methods.R
\docType{methods}
\name{quantileTrim}
\alias{quantileTrim}
\alias{quantileTrim,matrix,NULL_OR_missing-method}
\alias{quantileTrim,matrix,matrix-method}
\title{Advanced usage: Quantile trim a matrix or matrix pair}
\usage{
quantileTrim(mat1, mat2 = NULL, lower = 0, upper = 1)

\S4method{quantileTrim}{matrix,NULL_OR_missing}(mat1, mat2 = NULL,
  lower = 0, upper = 1)

\S4method{quantileTrim}{matrix,matrix}(mat1, mat2 = NULL, lower = 0,
  upper = 1)
}
\arguments{
\item{mat1}{matrix: Sites in rows, positions in columns.}

\item{mat2}{matrix or NULL: Optional second matrix with the same shape.}

\item{lower}{numeric: Lower cutoff for trimming (between 0-1).}

\item{upper}{numeric: Upper cutoff for trimming (between 0-1).}
}
\value{
trimmed matrix or list of trimmed matrices
}
\description{
Internal function for performing the quantile trimming. Removes outlier rows
with the highest or lowest sums. If two matrices are supplied, trims both
matrices simultaniously.
}
\examples{
# Make some random matrices
m1 <- replicate(5, rnorm(10))
m2 <- replicate(5, rnorm(10))

# Trim a single matrix
quantileTrim(mat1=m1, upper=0.75)

# Trim a pair of matrices (e.g. forward/reverse strand)
quantileTrim(mat1=m1, mat2=m2, upper=0.75)
}
