% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{relations}
\alias{relations}
\alias{starts_before_end_of}
\alias{ends_after_start_of}
\alias{starts_before_start_of}
\alias{starts_after_start_of}
\alias{ends_before_end_of}
\alias{ends_after_end_of}
\alias{ends_before_start_of}
\alias{starts_after_end_of}
\alias{meets}
\alias{met_by}
\alias{contemporary_with}
\alias{starts_during}
\alias{includes_start_of}
\alias{ends_during}
\alias{includes_end_of}
\alias{starts_with}
\alias{ends_with}
\alias{overlaps_before}
\alias{overlaps_after}
\alias{includes}
\alias{included_in}
\alias{begins}
\alias{begun_by}
\alias{ends}
\alias{ended_by}
\alias{equal_to}
\title{Temporal relations}
\usage{
starts_before_end_of(x, y, strict = FALSE)

ends_after_start_of(x, y, strict = FALSE)

starts_before_start_of(x, y, strict = FALSE)

starts_after_start_of(x, y, strict = FALSE)

ends_before_end_of(x, y, strict = FALSE)

ends_after_end_of(x, y, strict = FALSE)

ends_before_start_of(x, y, strict = FALSE)

starts_after_end_of(x, y, strict = FALSE)

meets(x, y, strict = FALSE)

met_by(x, y, strict = FALSE)

contemporary_with(x, y, strict = FALSE)

starts_during(x, y, strict = FALSE)

includes_start_of(x, y, strict = FALSE)

ends_during(x, y, strict = FALSE)

includes_end_of(x, y, strict = FALSE)

starts_with(x, y, strict = FALSE)

ends_with(x, y, strict = FALSE)

overlaps_before(x, y, strict = FALSE)

overlaps_after(x, y, strict = FALSE)

includes(x, y, strict = FALSE)

included_in(x, y, strict = FALSE)

begins(x, y, strict = FALSE)

begun_by(x, y, strict = FALSE)

ends(x, y, strict = FALSE)

ended_by(x, y, strict = FALSE)

equal_to(x, y, strict = FALSE)
}
\arguments{
\item{x, y}{Pair(s) of periods to test the relation between. Can be
\code{\link[=interval]{interval()}} objects, two-element numeric vectors, or lists of two-element
numeric vectors.}

\item{strict}{By default, comparison is inclusive (i.e. using \code{<=} and \code{>=}).
Use \code{strict = TRUE} for strict comparison (i.e. using \code{<} and \code{>}).}
}
\value{
Logical vector the same length as \code{x} and \code{y}.
}
\description{
These functions test for the logical relation between two periods according
to Levy's typology \insertCite{LevyEtAl2021,Levy2025}{tempo}.
}
\examples{
period1 <- c(1500, 1900)
period2 <- c(1800, 1950)
period3 <- c(1900, 1950)

contemporary_with(period1, period2)

# Inclusive relations (the default)
contemporary_with(period1, period3)

# Strict relations
contemporary_with(period1, period3, strict = TRUE)
}
\references{
\insertAllCited{}
}
