% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RleListList-class}
\alias{RleListList-class}
\alias{.RleListList}
\alias{RleListList}
\title{List of RleList-objects}
\usage{
RleListList(...)
}
\arguments{
\item{...}{RleLists}
}
\value{
RleListList
}
\description{
List of RleList-objects
}
