% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo-package.R
\docType{package}
\name{tempo-package}
\alias{tempo}
\alias{tempo-package}
\title{tempo: Chronological Intervals and Temporal Relations}
\description{
Provides a formal representation of intervals between two points in time (periods) and the potential logical relations between them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joeroe/tempo}
  \item Report bugs at \url{https://github.com/joeroe/tempo/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

}
\keyword{internal}
