% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMPO-package.R
\docType{package}
\name{TEMPO-package}
\alias{TEMPO-package}
\alias{TEMPO}
\title{R Tools for NSI Romania Open data}
\description{
Brief summary of the eurostat package
}
\details{
\tabular{ll}{
Package: \tab TEMPO\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2018\cr
License: \tab GPL3 + EUPL LICENSE\cr
LazyLoad: \tab yes\cr
}

R Tools for NSI Romania Open data
}
\examples{
library(TEMPO)
}
\references{
See citation("TEMPO")
\url{https://github.com/MarianNecula/TEMPO}
}
\author{
Marian Necula, Bogdan Oancea, Ana Maria Tiru \email{marian.necula@insse.ro} \url{https://github.com/MarianNecula/TEMPO}
}
\keyword{package}
