% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_bulk.R
\name{tempo_bulk}
\alias{tempo_bulk}
\title{Download one or multiple tables from TEMPO Online database}
\usage{
tempo_bulk(codes = list(), language = NULL, directory = NULL)
}
\arguments{
\item{codes}{- a list containing one or more strings containing the code for
the table/matrix in TEMPO Online database. See more on
how to obtain a matrix code from \code{\link{tempo_toc}}}

\item{language}{- a string to set the language for the downloaded
tables. Options: "ro" - for Romanian and "en" - for English. If no parameter
is given, implicitly downloads tables in Romanian}

\item{directory}{- a valid path where the files will be downloaded. The implicit value 
is NULL, i.e. current working directory.}
}
\description{
tempo_bulk
TEMPO Online database bulk download method. The function checks the
directory supplied as parameter or the working directory if has named
files as TEMPO Online matrices. If those files are more recent than
the last date upon which the TEMPO matrices have been updated,
the function skips them at download.
}
\details{

}
\examples{
\dontrun{
# Download a single table and write on disk
tempo_bulk("SOM101D", "ro")

# Downloads the whole TEMPO db and write each table on disk.
# Can take a long time.
# tempo_bulk(TOC_TEMPO$code, "ro") 
}
}
