% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_geo_codes.R
\docType{data}
\name{tempo_geo_codes}
\alias{tempo_geo_codes}
\title{Dataset contains TEMPO Online geocodes.}
\format{A \linkS4class{list} with 5 \linkS4class{data.frame}.}
\usage{
tempo_geo_codes
}
\description{
This dataset provides the codes from the NUTS classification system, the full names and 
the codes used for retrieving data from TEMPO Online database service, for all 
macroregions, regions and counties from Romania. For macroregions and regions the package
provides dataframes with names in English.
}
\keyword{datasets}
