% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_get.R
\name{tempo_get}
\alias{tempo_get}
\title{Download a table from TEMPO Online database}
\usage{
tempo_get(matrix_code = NULL, language = NULL)
}
\arguments{
\item{matrix_code}{- string containing the code for 
the table/matrix in TEMPO Online database. See more on
how to obtain a matrix code from \code{\link{tempo_codes}}}
}
\value{
Returns a dataframe object.
}
\description{
tempo_get downloads a a table from 
TEMPO Online database
}
\details{
This function sends GET/POST requests to TEMPO Online JSON service
and parses the content of the responses, using  \code{\link{httr}} wrappers
for \code{\link{RCurl::httpGET}} and \code{\link{RCurl::httpPOST}}.
The content of the \code{\link{httr::POST}} requests is parsed into a dataframe.
}
\examples{
\dontrun '{'tempo_get("SOM101D", "ro")'}'
}
\author{
Bogdan Oancea, Ana Tiru and Marian Necula 
members of the Romanian NIS Experimental Statistics
\url{http://www.insse.ro/cms/ro/statistici-experimentale}
\email{statistici.experimentale@insse.ro}
}
