% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_toc.R
\name{tempo_toc}
\alias{tempo_toc}
\title{tempo_toc}
\usage{
tempo_toc(fullDescription = FALSE)
}
\arguments{
\item{full_description}{- implicitly set as false
if fullDescription == TRUE then starts collecting dates for last updates}
}
\value{
Returns a dataframe object.
}
\description{
Get TOC for Tempo Online
function gets a parameter full_description implicitly false 
if fullDescription == TRUE then starts collecting dates for last updates
}
\details{
This functions sends one or multiple GET requests and parses the content into 
a dataframe
}
\examples{
tempo_toc(fullDescription = FALSE)

}
